Layout 프로토콜 리팩토링
