+++ Sources/SwiftLayout/Constraint/Extension/UIView+Constraint.swift
@@ -10,7 +10,7 @@ import UIKit
 
 extension Layout where Self: UIView {
     
-    public func constraint<C>(@ConstraintBuilder _ content: () -> C) -> ConstraintLayout<Self, C> where C: Constraint {
+    public func constraint<C>(@ConstraintBuilder _ content: () -> C) -> ConstraintLayout<C> where C: Constraint {
         .init(view: self, constraint: content())
     }
