+++ Sources/SwiftLayout/Layout/Extension/UIView+Layout.swift
@@ -2,29 +2,25 @@
 //  UIView+Layout.swift
 //  
 //
-//  Created by oozoofrog on 2022/02/01.
+//  Created by oozoofrog on 2022/02/06.
 //
 
 import Foundation
 import UIKit
 
-extension UIView: LayoutAttachable {
+extension Layout where Self: UIView {
     
-    public func active() -> AnyDeactivatable {
-        return AnyDeactivatable()
+    public func callAsFunction<L>(@LayoutBuilder _ build: () -> L) -> ViewLayout<L> where L: Layout {
+        ViewLayout(view: self, layoutable: build())
     }
     
-    public func deactive() {
+    public func attachSuperview(_ superview: UIView?) {
+        superview?.addSubview(self)
+    }
+    
+    public func detachFromSuperview(_ superview: UIView?) {
+        guard self.superview == superview else { return }
         removeFromSuperview()
-    }
-    
-    public func attachLayout(_ layout: LayoutAttachable) {
-        translatesAutoresizingMaskIntoConstraints = false
-        layout.addSubview(self)
-    }
-    
-    public func constraints(with view: UIView) -> [NSLayoutConstraint] {
-        []
     }
     
     public var hashable: AnyHashable {
@@ -32,12 +28,5 @@ extension UIView: LayoutAttachable {
     }
 }
 
-public extension Layout where Self: UIView {
-    
-    @discardableResult
-    func callAsFunction<Sub>(@LayoutBuilder _ content: () -> Sub) -> SuperSubLayout<Self, Sub> where Sub: Layout {
-        SuperSubLayout<Self, Sub>(superview: self, subLayout: content())
-    }
-    
-}
+extension UIView: Layout {}
