+++ Sources/SwiftLayout/Layout/Layouts/ViewLayout.swift
@@ -1,0 +1,18 @@
+//
+//  ViewLayout.swift
+//  
+//
+//  Created by oozoofrog on 2022/02/06.
+//
+
+import Foundation
+import UIKit
+
+public struct ViewLayout<L>: LayoutViewContainable where L: LayoutContainable {
+    var view: UIView
+    var layoutable: L
+    
+    public var layouts: [Layout] {
+        layoutable.layouts
+    }
+}
